pyproject.toml
[project]
name = "mobile_parser"
version = "0.1.0"
description = "Parser of mobile API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl==3.1.5",
    "pandas==3.0.1",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "urllib3>=2.6.3",
    "tqdm>=4.67.3"
]

[project.scripts]
parser = "main:__main__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "black==26.1.0",
    "mypy==1.19.1",
    "types-requests==2.32.4.20260107"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_optional = true
ignore_missing_imports = true
follow_imports = "skip"

